feat: add new methods in VoucherService and refine validation exception handling
- Added several new methods to VoucherService
- Updated handleMethodArgumentNotValidException to separate field errors and global errors for clearer response structure

+++ e-commerce/src/main/java/com/e_commerce/e_commerce/dto/response/VoucherResponse.java
@@ -16,9 +16,9 @@ public class VoucherResponse {
 
     String code;
 
-    BigDecimal discountAmount;
+    Integer discountAmount;
 
-    BigDecimal discountPercent;
+    Integer discountPercent;
 
     LocalDateTime validFrom;
