+++ e-commerce/src/main/java/com/e_commerce/e_commerce/entity/Voucher.java
@@ -24,9 +24,9 @@ public class Voucher {
 
     String code;
 
-    BigDecimal discountAmount;
+    Integer discountAmount;
 
-    BigDecimal discountPercent;
+    Integer discountPercent;
 
     LocalDateTime validFrom;
